i32

+++ backends/gstreamer/media_stream.rs
@@ -227,8 +227,8 @@ impl GStreamerMediaStream {
             let caps = gst::Caps::builder("video/x-raw")
                 .field("format", &gst_video::VideoFormat::Bgra.to_string())
                 .field("pixel-aspect-ratio", &gst::Fraction::from((1, 1)))
-                .field("width", &size.width)
-                .field("height", &size.height)
+                .field("width", &(size.width as i32))
+                .field("height", &(size.height as i32))
                 .build();
             source
                 .set_property("caps", &caps)
